+++ src/Components/UnkownGuessMsg/UnknownGuessMsg.tsx
@@ -1,7 +1,6 @@
 import './unknownGuess.css';
 
 export default function UnknownGuessMsg({ inputValue }: {inputValue: string} ) {
-    console.log(inputValue);
     return (
         <div className="unknown-guess-msg">
             Neznana občina: { inputValue }