+++ src/Components/WrongGuessMsg/WrongGuessMsg.tsx
@@ -3,7 +3,7 @@ import './wrongGuessMsg.css';
 export default function WrongGuessMsg() {
     return (
         <div className="wrong-guess-msg">
-            Wrong guess! Try again.
+            Napačen odgovor
         </div>
     )
 }